Splits python representation of metadata and its parser
The current implementation of python representation of metadata and metadata parser was tightly interconnected. Support for other versions of OData was not possible as in each version elements are added, removed or modified. Therefore, we decided to split metadata representation and its parser. With this approach, we can easily define supported elements and its parsing functions in a single class. This "configuration" class is stateless and has to be a child of ODATAVersion.

Additional changes including updating directory structure and refactoring old code to accommodate for incoming ODATA V4 support.

New module model:
- builder -> MetadataBuilder was moved here to make code easier to read
elements -> All EDM elements were moved here, to make python representation of elements version independent. All parsable elements have to inherit from "from_etree_mixin".

- from_etree_callbacks -> All  from_etree static methods were moved into separated function. This is a naive approach as its premise is that all from_etree implementations will be reusable in version V4.

- types_traits -> "types traits" were moved here to make code cleaner and easier to read.

Module V2:
- __init__ -> includes OData2 definition.

- service -> function-wise nothing has been changed.

"Main" module:
- config -> class Config was moved here to make it version and model-independent. In case we will ever need a config class also for service. Also ODataVersion class lives here.

- policies -> All policies were moved here as well as ParserError enum. Again to make policies version and model-independent.

Tests were only updated to incorporate new API.
