+++ pyodata/client.py
@@ -3,9 +3,11 @@
 import logging
 import warnings
 
-import pyodata.v2.model
-import pyodata.v2.service
+from pyodata.config import Config
+from pyodata.model.builder import MetadataBuilder
 from pyodata.exceptions import PyODataException, HttpError
+from pyodata.v2.service import Service
+from pyodata.v2 import ODataV2
 
 
 def _fetch_metadata(connection, url, logger):
@@ -34,43 +36,37 @@ class Client:
     """OData service client"""
 
     # pylint: disable=too-few-public-methods
-
-    ODATA_VERSION_2 = 2
-
-    def __new__(cls, url, connection, odata_version=ODATA_VERSION_2, namespaces=None,
-                config: pyodata.v2.model.Config = None, metadata: str = None):
+    def __new__(cls, url, connection, namespaces=None,
+                config: Config = None, metadata: str = None):
         """Create instance of the OData Client for given URL"""
 
         logger = logging.getLogger('pyodata.client')
 
-        if odata_version == Client.ODATA_VERSION_2:
+        # sanitize url
+        url = url.rstrip('/') + '/'
 
-            # sanitize url
-            url = url.rstrip('/') + '/'
+        if metadata is None:
+            metadata = _fetch_metadata(connection, url, logger)
+        else:
+            logger.info('Using static metadata')
 
-            if metadata is None:
-                metadata = _fetch_metadata(connection, url, logger)
-            else:
-                logger.info('Using static metadata')
+        if config is not None and namespaces is not None:
+            raise PyODataException('You cannot pass namespaces and config at the same time')
 
-            if config is not None and namespaces is not None:
-                raise PyODataException('You cannot pass namespaces and config at the same time')
+        if config is None:
+            logger.info('No OData version has been provided. Client defaulted to OData v2')
+            config = Config(ODataV2)
 
-            if config is None:
-                config = pyodata.v2.model.Config()
+        if namespaces is not None:
+            warnings.warn("Passing namespaces directly is deprecated. Use class Config instead", DeprecationWarning)
+            config.namespaces = namespaces
 
-            if namespaces is not None:
-                warnings.warn("Passing namespaces directly is deprecated. Use class Config instead", DeprecationWarning)
-                config.namespaces = namespaces
+        # create model instance from received metadata
+        logger.info('Creating OData Schema (version: %d)', config.odata_version)
+        schema = MetadataBuilder(metadata, config=config).build()
 
-            # create model instance from received metadata
-            logger.info('Creating OData Schema (version: %d)', odata_version)
-            schema = pyodata.v2.model.MetadataBuilder(metadata, config=config).build()
+        # create service instance based on model we have
+        logger.info('Creating OData Service (version: %d)', config.odata_version)
+        service = Service(url, schema, connection)
 
-            # create service instance based on model we have
-            logger.info('Creating OData Service (version: %d)', odata_version)
-            service = pyodata.v2.service.Service(url, schema, connection)
-
-            return service
-
-        raise PyODataException('No implementation for selected odata version {}'.format(odata_version))
+        return service
