+++ tests/test_client.py
@@ -1,25 +1,20 @@
 """PyOData Client tests"""
 
+from unittest.mock import patch
 import responses
 import requests
 import pytest
+
 import pyodata
-import pyodata.v2.service
-from unittest.mock import patch
+
 from pyodata.exceptions import PyODataException, HttpError
-from pyodata.v2.model import ParserError, PolicyWarning, PolicyFatal, PolicyIgnore, Config
+from pyodata.policies import ParserError, PolicyWarning, PolicyFatal, PolicyIgnore
+from pyodata.config import Config
+
+from pyodata.v2.service import Service
+from pyodata.v2 import ODataV2
 
 SERVICE_URL = 'http://example.com'
-
-
-@responses.activate
-def test_invalid_odata_version():
-    """Check handling of request for invalid OData version implementation"""
-
-    with pytest.raises(PyODataException) as e_info:
-        pyodata.Client(SERVICE_URL, requests, 'INVALID VERSION')
-
-    assert str(e_info.value).startswith('No implementation for selected odata version')
 
 
 @responses.activate
@@ -46,13 +41,13 @@ def test_create_service_application_xml(metadata):
 
     client = pyodata.Client(SERVICE_URL, requests)
 
-    assert isinstance(client, pyodata.v2.service.Service)
+    assert isinstance(client, Service)
 
     # onw more test for '/' terminated url
 
     client = pyodata.Client(SERVICE_URL + '/', requests)
 
-    assert isinstance(client, pyodata.v2.service.Service)
+    assert isinstance(client, Service)
 
 
 @responses.activate
@@ -68,13 +63,13 @@ def test_create_service_text_xml(metadata):
 
     client = pyodata.Client(SERVICE_URL, requests)
 
-    assert isinstance(client, pyodata.v2.service.Service)
+    assert isinstance(client, Service)
 
     # onw more test for '/' terminated url
 
     client = pyodata.Client(SERVICE_URL + '/', requests)
 
-    assert isinstance(client, pyodata.v2.service.Service)
+    assert isinstance(client, Service)
 
 
 @responses.activate
@@ -127,6 +122,7 @@ def test_client_custom_configuration(mock_warning, metadata):
     }
 
     custom_config = Config(
+        ODataV2,
         xml_namespaces=namespaces,
         default_error_policy=PolicyFatal(),
         custom_error_policies={
@@ -145,10 +141,10 @@ def test_client_custom_configuration(mock_warning, metadata):
         'Passing namespaces directly is deprecated. Use class Config instead',
         DeprecationWarning
     )
-    assert isinstance(client, pyodata.v2.service.Service)
+    assert isinstance(client, Service)
     assert client.schema.config.namespaces == namespaces
 
     client = pyodata.Client(SERVICE_URL, requests, config=custom_config)
 
-    assert isinstance(client, pyodata.v2.service.Service)
+    assert isinstance(client, Service)
     assert client.schema.config == custom_config
